:sparkles: [#20] Map과 NearByMap 분리
