+++ components/Map/Map.tsx
@@ -1,40 +1,12 @@
 'use client'
-import React, { useEffect, useRef, useState } from 'react'
-import useGeolocation from 'react-hook-geolocation'
-import { useGetBusStationAroundListOut } from '@/lib/apis/bus-station-service/api/bus-station-service'
+import React, { useEffect, useRef } from 'react'
 
-const Map = () => {
-  const [map, setMap] = useState<naver.maps.Map>()
-  const [current, setCurrent] = useState<naver.maps.Marker>()
-  const nearBys = useRef<naver.maps.Marker[]>()
-  const [location, setLocation] = useState<{ lat: number; lng: number } | undefined>(undefined)
+interface MapProps {
+  setMap: (map?: naver.maps.Map) => void
+  className: string
+}
+function Map({ setMap, className }: MapProps) {
   const mapElement = useRef(null)
-  const geolocation = useGeolocation()
-  const { data } = useGetBusStationAroundListOut(
-    { x: location?.lng || 0, y: location?.lat || 0 },
-    location?.lng !== undefined && location?.lat !== undefined,
-  )
-
-  useEffect(() => {
-    if (geolocation.latitude & geolocation.longitude) {
-      const currentLocation = {
-        lat: geolocation.latitude,
-        lng: geolocation.longitude,
-      }
-      setLocation(currentLocation)
-
-      const marker = new naver.maps.Marker({
-        position: currentLocation,
-        map,
-      })
-
-      map?.morph(currentLocation)
-      setCurrent(marker)
-    }
-    return () => {
-      current?.setMap(null)
-    }
-  }, [geolocation.latitude])
 
   useEffect(() => {
     const { naver } = window
@@ -50,29 +22,11 @@ const Map = () => {
     setMap(map)
 
     return () => {
-      current?.setMap(null)
+      setMap(undefined)
     }
   }, [])
 
-  useEffect(() => {
-    const { naver } = window
-    if (!mapElement.current || !naver) return
-
-    const markers = data?.map(
-      (item) =>
-        new naver.maps.Marker({
-          map: map,
-          position: new naver.maps.LatLng(item.y, item.x),
-        }),
-    )
-
-    nearBys.current = markers
-    return () => {
-      nearBys.current?.forEach((marker) => marker.setMap(null))
-    }
-  }, [data])
-
-  return <div ref={mapElement} id="map" className="w-full h-300" />
+  return <div ref={mapElement} id="map" className={className} />
 }
 
 export default Map
